compare with dopex

+++ contracts/compare/partlikeits1983/AdapterParty.sol
@@ -12,8 +12,7 @@ contract AdapterParty {
         uint64 volatility,
         uint64 rate
     ) external view returns (uint256 price, uint256 gasUsed) {
-        int256 call;
-        // uint256 put;
+        uint256 call;
         uint256 startGas;
         uint256 endGas;
 
@@ -27,21 +26,16 @@ contract AdapterParty {
             int256(int64(volatility))
         );
 
-        // int128 spot64x64 = int128(uint128(uint256(spot) * 2 ** 64 / 1e18));
-        // int128 strike64x64 = int128(uint128(uint256(strike) * 2 ** 64 / 1e18));
-        // int128 timeToExpiry64x64 = int128(uint128((uint256(timeToExpirySec) * 1e18 / (365 * 24 * 60 * 60)) * 2 ** 64 / 1e18));
-        // int128 volatility64x64 = int128(uint128(uint256(volatility) * 2 ** 64 / 1e18));
-
         startGas = gasleft();
 
         int result = BS.c_BS_CALL(inputs);
         if (result > 0) {
-            call = result;
+            call = uint256(result);
         }
 
         endGas = gasleft();
         
         gasUsed = startGas - endGas;
-        return (uint256(call), gasUsed);
+        return (call, gasUsed);
     }
 }
